Refactor shiftSectioningLevel into its own class
