+++ src/commander.ts
@@ -585,105 +585,7 @@ export class Commander {
      * @param change
      */
     shiftSectioningLevel(change: 'promote' | 'demote') {
-        if (change !== 'promote' && change !== 'demote') {
-            throw TypeError(
-            `Invalid value of function parameter 'change' (=${change})`
-            )
-        }
-
-        const editor = vscode.window.activeTextEditor
-        if (editor === undefined) {
-            return
-        }
-
-        const promotes = {
-            part: 'part',
-            chapter: 'part',
-            section: 'chapter',
-            subsection: 'section',
-            subsubsection: 'subsection',
-            paragraph: 'subsubsection',
-            subparagraph: 'paragraph'
-        }
-        const demotes = {
-            part: 'chapter',
-            chapter: 'section',
-            section: 'subsection',
-            subsection: 'subsubsection',
-            subsubsection: 'paragraph',
-            paragraph: 'subparagraph',
-            subparagraph: 'subparagraph'
-        }
-
-        function replacer(
-            _match: string,
-            sectionName: keyof typeof promotes,
-            asterisk: string | undefined,
-            options: string | undefined,
-            contents: string
-        ) {
-            if (change === 'promote') {
-                return '\\' + promotes[sectionName] + (asterisk ?? '') + (options ?? '') + contents
-            } else {
-                // if (change === 'demote')
-                return '\\' + demotes[sectionName] + (asterisk ?? '') + (options ?? '') + contents
-            }
-        }
-
-        // when supported, negative lookbehind at start would be nice --- (?<!\\)
-        const pattern = /\\(part|chapter|section|subsection|subsection|subsubsection|paragraph|subparagraph)(\*)?(\[.+?\])?(\{.*?\})/g
-
-        function getLastLineLength(someText: string) {
-            const lines = someText.split(/\n/)
-            return lines.slice(lines.length - 1, lines.length)[0].length
-        }
-
-        const document = editor.document
-        const selections = editor.selections
-        const newSelections: vscode.Selection[] = []
-
-        const edit = new vscode.WorkspaceEdit()
-
-        for (let selection of selections) {
-            let mode: 'selection' | 'cursor' = 'selection'
-            let oldSelection: vscode.Selection | null = null
-            if (selection.isEmpty) {
-                mode = 'cursor'
-                oldSelection = selection
-                const line = document.lineAt(selection.anchor)
-                selection = new vscode.Selection(line.range.start, line.range.end)
-            }
-
-            const selectionText = document.getText(selection)
-            const newText = selectionText.replace(pattern, replacer)
-            edit.replace(document.uri, selection, newText)
-
-            const changeInEndCharacterPosition = getLastLineLength(newText) - getLastLineLength(selectionText)
-            if (mode === 'selection') {
-                newSelections.push(
-                    new vscode.Selection(selection.start,
-                        new vscode.Position(selection.end.line,
-                            selection.end.character + changeInEndCharacterPosition
-                        )
-                    )
-                )
-            } else if (oldSelection) { // mode === 'cursor'
-                const anchorPosition = oldSelection.anchor.character + changeInEndCharacterPosition
-                const activePosition = oldSelection.active.character + changeInEndCharacterPosition
-                newSelections.push(
-                    new vscode.Selection(
-                        new vscode.Position(oldSelection.anchor.line, anchorPosition < 0 ? 0 : anchorPosition),
-                        new vscode.Position(oldSelection.active.line, activePosition < 0 ? 0 : activePosition)
-                    )
-                )
-            }
-        }
-
-        vscode.workspace.applyEdit(edit).then(success => {
-            if (success) {
-                editor.selections = newSelections
-            }
-        })
+       this.extension.section.shiftSectioningLevel(change)
     }
 
     devParseLog() {
